src/scripts/requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
duckdb==0.8.1
filelock==3.12.2
fsspec==2023.6.0
huggingface-hub==0.15.1
idna==3.4
packaging==23.1
python-dotenv==1.0.0
PyYAML==6.0
requests==2.31.0
tdqm==0.0.1
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==2.0.3
